For development purposes, add a MediaWiki API proxy
For local development, we are prevented by CORS from making requests
directly to the MediaWiki API; there are various ways around this, but
what I've done in this commit is to create a proxy in the
verification-pages web app which proxies requests to the MediaWiki API.

The proxy uses the mediawiki_api Ruby gem with the Wikidata login
credentials from the environment.

There's some mismatch between these interfaces; the main one is that the
mediawiki_api handles getting the CSRF token for requests on our behalf,
whereas in the Javascript running on www.wikidata.org or
test.wikidata.org we would have to get that token and supply it as
necessary.