Handle people being merged into someone else after reconciliation
Prior to this commit, a statement that's been successfully actioned and
goes to "done" will move back to "actionable" if they're later merged
into someone else (which effectively deletes the original person we
reconciled and against, and redirects them to the person they were
merged into).

This commit changes the query we make to find current members of the
legislature to include previous item IDs the people might have had in
the past, and will match the person based on those item IDs as well as
the person's current item ID when looking them up.

This should fix a good proportion of the cases in #105, where people are
still in "actionable" although the have the right P39 already.