Record the last time a statement was actioned on Statement#actioned_at
We are going to use this new actioned_at field to support a new
"reverted" state, which detects when we have enacted a suggestion (i.e.
taken it from "actionable" -> "done") but that position held statement
was later changed in Wikidata so it's no longer detected as "done".

This commit, however, should just keep the current behaviour the same as
before, but relying on the actioned_at column of the statements table,
instead of keys with a 5 minute timeout in the Rails cache.