Add "reverted" for where actioned statements no longer match Wikidata
This is a new state to represent the fairly common situation that somone
has actioned a statement (that is: enacted it in Wikidata, moving the
statement to "done") but subsequenty it doesn't match what's in
Wikidata. This might be due to a Wikidata user reverting the change (or
a bug in our software) but in any case, we need to make sure we don't
encourage someone to try to action the satement again, which is what
would happen if it went back into the actionable state - hence this new
state instead, which doesn't prompt someone to make a change.

Note that to account for caching delays in detecting what's changed in
Wikidata, a statement still goes to "done" for 5 minutes after
actionining, regardless of whether it matches what we're getting back
from Wikidata queries - if they don't matched Wikidata more than 5
minutes after actioning, they will go to "reverted" however.