Generate error messages for each problem implying 'manually_actionable'
We want to move towards exposing in the interface the reasons *why* a
statement is in the manually_actionable state. This commit doesn't
expose these problems in the interface yet, but does switch how the
manually_actionable state is calculated to check whether the list of
messages describing problems is empty.