Don't use ENV.fetch for WikiClient config
Using ENV.fetch causes an exception to be raised, which means that when
these settings aren't set (e.g. when running the tests on Travis) an
exception is raised.

For now I've just done the simplest thing and used the #[] method rather
than #fetch to prevent the exception from being raised.

In the future we might want to think about adding extra checks to ensure
these are provided, but run the checks in methods on an instance, rather
than in the module definition like it was previously.