Loading statements no longer needs to find the country from Wikidata
Each Page is now associated with a Country, which has a country code -
this means that the RetrieveCountryPosition service is no longer needed.
This is worth changing since the query would fail to find a country for
some positions, so it should be more reliable removing this, as well as
making fewer requests to query.wikidata.org.

The other information retrieved by the service's query, namely the
position in the 'subclass of' hierarchy, wasn't used by the calling
code.

There's a change to one of the tests as well, which previously
inconsistently returned Italian data from the query run by
RetrieveCountryPosition even though the Page under test was associated
with a Country of Canada.