Use the term in the RetrievePositionData query
If we have a verification page with a term and there are existing
positions on Wikidata with start dates we can ignore any if the term
didn't start before the position.

This means we never have the old position statement from a previous
term. Instead the app will have to create a new position statement when
the statement is actioned.

Example query: http://tinyurl.com/y7onjruh