RetrievePositionData: make the query method public
We'd like to be able to make it easy to run the query for finding
existing position in the Wikidata Query Service; but previously you
couldn't get the query that'll be used without running it. (Also the
existing private query method only returned the templated query.)

This commit factors out a public method that returns the query that
would actually be run.