Add a link to try a page's SPARQL query on WQS
For debugging problems, it's really helpful to be able to easily run the
SPARQL query that's used for finding existing position held statements
in Wikidata. This commit adds a link to the 'show' page for pages in the
admin interface that'll load up the query that's used into the
Wikidata Query Service (WQS).