Handle a term having both P580 (start time) and P571 (inception)
If a term has both a 'start time' and an 'inception', at the moment we
return two rows for each P39. Instead we should just pick one value.
This could be done with a relatively simple COALESCE if we always prefer
'inception', but it's useful to us to have term dates precise to the
day, so if 'start time' is of higher precision, we use that. Generally
we prefer 'inception', however, since that's the Wikidata consensus for
how the start of a term should be represented.

Fixes #239