Reconciliation searches Wikidata and Wikipedia in parallel
The two searches are completely independent, and we’re already inside a
Promise, so it’s super easy to run the searches in parallel, rather than
series, and then resolve the Promise once they’re both complete.

This saves time. In my development environment, "searching" went down
from ~5 seconds to ~3 seconds.

It was also a good opportunity to break up the monolithic search method
in wikiapi.js. Hopefully now it’s a bit easier to follow.