Skip verification of non-suggestions-store data that's already correct
When data comes from suggestions-store, it's important for us to check
whether it's correct, regardless of whether Wikidata already has that
information, so that our partner have a signal about whether individual
crowd-sourcers are performing well or not, and because we only allow
data through to their build which has been verified.

However, for external sources, where we're just using a verification
page to make sure the country is up-to-date, you usually would want to
completely skip over those cases since nothing needs to be done,
except possibly adding a reference URL.

Tony Bowden pointed out that we might want this to be a new state
("matches but not checked") so that people can go through those to add
the reference URLs, but will require updating the UI, and this is a good
first step towards that. (Essentially the same elsif clause can be used
to define that new state.)

Fixes #181