Refactor StatementDecorator#done? method
I want to update the electoral district check to check whether we're
dealing with an executive position, but I didn't want to add further
complexity to the `done?` method. So I've pulled out the checks for each
bit of data we're interested in into separate methods to make changing
them simpler.