Ignore district if it's not provided in the source
If a statement doesn't have a district then ignore that field when
comparing to Wikidata.  This should prevent statements going into the
`manually_reconcilable` state when the source CSV row is missing
district but it's there in Wikidata.

For now I'm only handling district. We probably want to make a similar
change for group and term, at least, but for now this is the minimal
change possible to take a step in that direction.

Ultimately it might be good to have a more generic solution, so if we
decide to add more fields in the future then they'll automatically be
handled, but for now I think this is OK.