Don't let LoadStatements wipe out reconciliations people have done
If someone does a manual reconciliation on a Verification Page, then
that sets the person_item on that Statement. However, when the cron job
runs to update statements from the CSV source, the blank value from the
CSV file would overwrite person_item, losing the good work they'd done
and putting the statement into a confusing state. (We're also moving to
reconciling parliamentary groups and electoral districts, so the same
concern applies the reconciled values for them too.)

This changes LoadStatements.run so that it only updates *_item
attributes if there's actually a value for them in the CSV file.

Partial fix for #305