Add a rake task to restore lost reconciled item IDs
We had a bug that meant that reconciled item IDs that people had found
on the verification page were being wiped out when reloading the
statements once a day. Fortunately the reconciliations are also logged
in the reconciliations table. This Rake task will go through every
Reconciliation and trigger it to update the statement it refers to, so
we can use this to restore the missing values.