Add better error-checking to getPropertyID
If you supplied an unknown property ID to this function, the error you'd
get wouldn't be very helpful, because some completely different bit of
code would be using undefined instead of a property ID string.

The error if you provide an unknown server name could also be more
helpful.

This commit improves the error message in both of those cases.