Replace transaction ID with a link to the statement
Most users will never need to know the transaction ID, but it's still
useful to be able to get at it - we can kill two birds with one stone
by replacing it with a useful "copy link to statement" section marker.

This also adds styling so that when the table row is targetted it has a
yellow background which fades to white.

Note that this table styling won't work on first page load because the
table row isn't in the DOM yet.