Move cell labels to top of table
An effort to improve information density, as per #337.

In modern browsers, the table header will stick to the top of the
window as you scroll down the table. This required a slight change to
how we collapsed table borders (swapping `border-collapse: collapse`
for `border-spacing: 0`) because the `border-collapse` was resulting
in half-width (!!) borders when combined with `position: sticky`.