Allow reconciling from any state except "verifiable"
Previously, `this.statement.type` always reflected the statement type
on the server. But now, to allow people to skip back to fix bad
reconciliations, we needed to temporarily modify the statement’s type
client-side.

The only situation where we’re not allowing people to jump to
reconcilable is if the statement hasn’t been verified yet, because we
don’t want them to waste time reconciling a potentially false statement.