Always use the "unverifiable" state for verification status: false
It was previously the case that a statment would be classified as
"manually actionable" or "reverted" even if someone had said that
it was "unverifiable". This would have been rare, but may come up
more now, because we're allowing people to correct the verification
status after the other actions have been taken.