Have a single, simple button for the 'actionable' state
This text didn't appear much any more, but it still provides a lot of
information that is only helpful to a very small minority of users, and
makes it look confusing and intimidating to the rest. It wasn't even very
helpful for developers, since it just showed how the qualifiers it would
like a P39 to have, and nothing about what was there already.

Fixes the first part of #384