Don't automatically action statements directly after bulk reconciliation
This is a workaround to prevent the situation where a bulk
reconciliation happens, and immediately afterwards a large number of
statements are actioned in parallel - this is bad Wikimedia API
etiquette. Instead, we flag updates of statements that came from a bulk
reconciliation, and don't automatically action statements if that was
the case. Moving into any other state clears the flag.