Display full country name on statistics page
If we already know about a country, i.e. we have a row in the countries
table for it, then we can display that countries full name on the
statistics page. If the country doesn't exist in the DB yet then display
the country code and link to the page for creating the row in the DB.