Fix errors uncovered by eslint
Most of these are unused variables, which can safely be removed. There
was one case where a `data` variable had been declared twice, so I've
renamed the second instance of that variable `referenceData`. There was
also a problem in the reconcilable component where the `created` hook
was defined twice, so I've rolled those together into one.