Update reference URL in the frontend UI
Allows the user to change the reference URL for a statement. When the
change the URL it will ask them whether they want the changed URL to
apply to all statements, or just this one.

This is a bit more complex than I'd like, because the actual change of
URL happens in the `Verifiable` component, which then triggers an event
that the parent (`Statement`) is listening to, which in turn needs to
trigger an event on _its_ parent (`App`), which is the component that
owns the data that we want to change, in this case the page and a
statement. The alternative would be to call `this.$parent.$parent` and
trigger events that way, but that seems to be recommended against in the
Vue.js world, from what I've read, because it breaks encapsulation and
can cause the code to be very brittle. So this change includes a fair
bit of boilerplate, which is just wiring together child -> parent
communication using events.