Move from_suggestions_store? from Statement to Page
Whether a statement is from suggestions-store or not is actually
entirely dependent on the Page, so it makes sense to be able to ask the
Page about that. (We also need that for a new API endpoint for
politician-data to consume.)

This still delegates that method from Statement to Page according to the
Law of Demeter principle.