Add WikiPageTemplateTag class
This class is a wrapper around `MediaWiki::Page::ReplaceableContent`. It
turns the template tag parameters on the wikipage into parameters
suitable for feeding into an instance of the `Page` model. It also has
methods for replacing the content on a wiki page and generating a
redirect URL back to the original wiki page.