Fix creation of equivalence claims
Error introduced in c3eb88579310911850a54a5727af5ace3d4cd519

To check if an ActiveRecord instance has changed from an `after_commit`
callback we need to check the previous changes prior to the transaction
being committed.

This is because `changes_applied` is called [1] before our callback.
This moves the current changes in to the `@previously_changed` variable
[2] which is read by `*_previously_changed?` predicates.

[1] https://github.com/rails/rails/blob/6eaae62/activerecord/lib/active_record/attribute_methods/dirty.rb#L19-L20
[2] https://github.com/rails/rails/blob/6eaae62/activemodel/lib/active_model/dirty.rb#L146