Add new SPARQL query for retrieving positions
This adds a `NewRetrievePositionData` class that is an exact copy of the
current `RetrievePositionData` class, but with a simplified SPARQL
query.

The new SPARQL query doesn't do any kind of date comparison, it just
returns all matching positions so we can do the date comparison in Ruby.

Once we've got the rest of the pieces together for doing the date
comparison in Ruby we can replace the old `RetrievePositionData` class
with this new one.